% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIdent_functions.R
\name{clustHeatmap}
\alias{clustHeatmap}
\title{clustHeatmap: Generate Heatmap of Cluster Coefficients}
\usage{
clustHeatmap(scIdentObj, coef = "abs")
}
\arguments{
\item{scIdentObj}{The output object from the `scIdent` function.}

\item{coef}{A string indicating the type of coefficients to plot. Can be "abs" for absolute coefficients or "prop" for normalized coefficients. Default is "abs".}
}
\value{
A ggplot object representing the heatmap of cluster coefficients.
}
\description{
This function generates a heatmap of cluster coefficients (absolute or normalized) from the output of `scIdent`.
}
\details{
`clustHeatmap` creates a heatmap using ggplot2, showing the coefficients for each cluster. The function can plot either absolute or normalized coefficients.
}
\examples{
\dontrun{
clustHeatmap(scIdentObj, coef = "abs")
}

}
