% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{.MIXER}
\alias{.MIXER}
\title{.MIXER
this is the main core function, it should not be called by users}
\usage{
.MIXER(
  expressionMatrix,
  signatureMatrix,
  functionMixture,
  useCores = 1L,
  verbose = FALSE
)
}
\arguments{
\item{expressionMatrix}{a GxS gene expression matrix. with row names genes ID as in the signature Matrix. Genes in rows, samples in columns}

\item{signatureMatrix}{a NxL gene signature matrix for L cell types}

\item{functionMixture}{: the deconvolution function: cibersort, nu.svm.robust.RFE, ls.rfe.abbas (tiped without quotation marks)}

\item{useCores}{: (integer) the number of cpus to use.}
}
\description{
.MIXER
this is the main core function, it should not be called by users
}
