% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{GetCellTypes}
\alias{GetCellTypes}
\title{GetCellTypes
returns the Root Mean Square Error between y - XB, with unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)}
\usage{
GetCellTypes(obj, delta = 0)
}
\arguments{
\item{obj}{an object of class MIXTURE (see MIXTURE)}

\item{delta}{(float) default 0, the threshold value to define a coefficient as cero}
}
\value{
a logical SxL matrix with TRUE if the coefficients > delta , FALSE otherwise.
}
\description{
GetCellTypes
returns the Root Mean Square Error between y - XB, with unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)
}
