% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{GetCorrelation}
\alias{GetCorrelation}
\title{GetCorrelation
returns the Correlation \code{cor(Y,XB)} unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)}
\usage{
GetCorrelation(obj, type = c("proportion", "absolute"))
}
\arguments{
\item{obj}{an object of class MIXTURE (see MIXTURE)}

\item{type}{a character string "proportion" or "absolute". short word is allowed (default: "proportion")}
}
\value{
a SxK data matrix with K cell types (according to the signature matrix) and S subjects
}
\description{
GetCorrelation
returns the Correlation \code{cor(Y,XB)} unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)
}
