% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{GetMetrics}
\alias{GetMetrics}
\title{GetMetrix
returns the Root Mean Square Error between y - XB, with unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)}
\usage{
GetMetrics(
  obj,
  metric = c("all", "RMSE", "R"),
  type = c("proportion", "absolute")
)
}
\arguments{
\item{obj}{an object of class MIXTURE (see MIXTURE)}

\item{metric}{: (character) "all" all the metrics (type ignored), "RMSE": (see GetRMSE), "R": (see GetCorrelation)}

\item{type}{a character string "proportion" or "absolute". short word is allowed (default: "proportions")}
}
\value{
if "all" : ACCmetric matrix, otherwise see GetRMSE or GetCorrelation
}
\description{
GetMetrix
returns the Root Mean Square Error between y - XB, with unnormalized B>0 (absolute) or normalized \code{sum(B)==1} (proportions)
}
