% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{GetPvalues}
\alias{GetPvalues}
\title{GetPvalues
returns the p values matrix.}
\usage{
GetPvalues(obj)
}
\arguments{
\item{obj}{an object of class MIXTURE (see MIXTURE)}
}
\value{
the p value matrix
}
\description{
if not calculated, stop
}
