% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{GetUsedGenes}
\alias{GetUsedGenes}
\title{GetUsedGenes
return the amount of signature genes in common with the input data}
\usage{
GetUsedGenes(obj)
}
\arguments{
\item{obj}{the object provided by MIXTURE}
}
\value{
a vector of Gene identifiers (i.e GeneSymbol)
}
\description{
GetUsedGenes
return the amount of signature genes in common with the input data
}
