% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{LM22}
\format{
A data frame with 547 rows (genes) and 22 variables (cell types). The rownames contains the GeneSymbols and the colnames the cell types
}
\usage{
data(LM22)
}
\description{
A dataset containing the molecular signatureprofiles of 22 cell types. taken from 
[Newman et al.](http://www.nature.com/nmeth/journal/v12/n5/abs/nmeth.3337.html)
}
\details{
\itemize{
  \item B cells naive
  \item B cells memory
  \item Plasma cells
  \item T cells CD8
  \item T cells CD4 naive
  \item T cells CD4 memory resting
  \item T cells CD4 memory activated
  \item T cells follicular helper
  \item T cells regulatory (Tregs)
  \item T cells gamma delta
  \item NK cells resting
  \item NK cells activated
  \item Monocytes
  \item Macrophages M0
  \item Macrophages M1
  \item Macrophages M2
  \item Dendritic cells resting
  \item Dendritic cells activated
  \item Mast cells resting
  \item Mast cells activated
  \item Eosinophils
  \item Neutrophils
}
}
\references{
\url{https://cibersort.stanford.edu/}
}
\keyword{datasets}
