% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{LoadMixtureResultsFromExcel}
\alias{LoadMixtureResultsFromExcel}
\title{LoadMixtureResultsFromExcel
This function load a MIXTURE object from an excel file}
\usage{
LoadMixtureResultsFromExcel(path)
}
\arguments{
\item{path}{the file path to the xlsx file}
}
\description{
LoadMixtureResultsFromExcel
This function load a MIXTURE object from an excel file
}
\seealso{
\code{\link{SaveExcel}}
}
