% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{nu.svm.robust.RFE}
\alias{nu.svm.robust.RFE}
\title{nu.svm.robust.RFE
noise constrained & recursive feature extraction based support vector nu-regression
it will solve the y = X*B problem, providing B>0}
\usage{
nu.svm.robust.RFE(X, y, nu = c(0.25, 0.5, 0.75), minProp = 0.007, maxiter = 6)
}
\arguments{
\item{X}{the signature matrix, a NxL matrix of N genes and L cell lines}

\item{y}{a vector of bulk gene expression sample}

\item{nu}{(float) the nu value, default nu = c(0.25,0.5,0.75)}

\item{minProp}{float. noise threshold level}

\item{maxiter}{default 6}
}
\description{
nu.svm.robust.RFE
noise constrained & recursive feature extraction based support vector nu-regression
it will solve the y = X*B problem, providing B>0
}
\author{
Elmer A. Fernández
}
