% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIdent_functions.R
\name{PlotDimCoef}
\alias{PlotDimCoef}
\title{PlotDimCoef: Plot Dimensional Reduction with Cluster Coefficients}
\usage{
PlotDimCoef(
  SeuratObj,
  scIdentObj,
  ms_celltypes = NULL,
  reduction,
  coef = "abs",
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{SeuratObj}{A Seurat object containing single-cell RNA-seq data.}

\item{scIdentObj}{The output object from the `scIdent` function, containing deconvolution results.}

\item{ms_celltypes}{A vector of cell types from the molecular signature to plot. Default is `NULL`, which plots all identified cell types.}

\item{reduction}{The type of dimensional reduction to use (e.g., "pca", "tsne", "umap").}

\item{coef}{A string indicating the type of coefficients to plot. Can be "abs" for absolute coefficients or "prop" for normalized coefficients. Default is "abs".}

\item{ncol}{Number of columns for the facet wrap. Default is `NULL`.}

\item{nrow}{Number of rows for the facet wrap. Default is `NULL`.}
}
\value{
Plots of dimensional reduction with cluster coefficients for each celltype detected by MIXTURE.
}
\description{
This function plots dimensional reduction on a Seurat object with cluster coefficients (absolute or normalized).
}
