% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{ProportionPlot}
\alias{ProportionPlot}
\title{ProportionPlot
a bar plot for each subject displaying the cell type proportions}
\usage{
ProportionPlot(obj)
}
\arguments{
\item{obj}{a MIXTURE object, see \code{\link{LoadMixtureResultsFromExcel}}}
}
\description{
ProportionPlot
a bar plot for each subject displaying the cell type proportions
}
\seealso{
\code{\link{LoadMixtureResultsFromExcel}}
}
