% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{SaveExcel}
\alias{SaveExcel}
\title{SaveExcel
It saves MIXTURE results as an Excel xlsx file}
\usage{
SaveExcel(obj, fileSave)
}
\arguments{
\item{obj}{the object provided by MIXTURE}

\item{fileSave}{(character) the xlsx file name}
}
\description{
It will build an xlsx file with named sheets as:
Absolute (with the absolute coefficients)
Proportions
Metrics
Pvalues
UsedGenes
}
\seealso{
\code{\link{LoadMixtureResultsFromExcel}}
}
