% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIdent_functions.R
\name{scIdent_MIXTURE}
\alias{scIdent_MIXTURE}
\title{scIdent_MIXTURE}
\usage{
scIdent_MIXTURE(X, y, nu = c(0.25, 0.5, 0.75), minProp = 0.001, maxiter = 6)
}
\arguments{
\item{X}{the signature matrix, a NxL matrix of N genes and L cell lines}

\item{y}{a vector of bulk gene expression sample}

\item{nu}{(float) the nu value, default nu = c(0.25,0.5,0.75)}

\item{minProp}{float. noise threshold level}

\item{maxiter}{default 6}
}
\description{
scIdent_MIXTURE
}
