% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureTest.R
\name{SelfTest}
\alias{SelfTest}
\title{SeflTest 
This function performe the Self Test, which means that the model should estimate
as pure cell types each cell type (column) present in the signature matrix}
\usage{
SelfTest(signatureMatrix)
}
\arguments{
\item{signatureMatrix}{a NxK molecular signature}
}
\value{
an MIXTURE object
}
\description{
SeflTest 
This function performe the Self Test, which means that the model should estimate
as pure cell types each cell type (column) present in the signature matrix
}
