% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureTest.R
\name{SimulatedMixtures}
\alias{SimulatedMixtures}
\title{SimulatedMixtures}
\usage{
SimulatedMixtures(signatureMatrix, maxCoefs, maxSamples, noisy = TRUE)
}
\arguments{
\item{signatureMatrix}{the NxK molecular signature matrix}

\item{maxCoefs}{(integer) the maximum number of simulated coefficients > 0}

\item{maxSamples}{(integer) the maximum number of simulated samples}

\item{noisy}{(logical) if TRUE, noisy samples will be provided}
}
\value{
a list with a simulated bulk tissues gene expression matrix nrow(signatureMatrix) by maxCoefs columns and their corresponding Beta coefficient matrix.
}
\description{
the Beta coefficnets are drown from a uniform distribution between 0.2 to 1
}
\details{
It returns the simulated MIxtures from the given molecular signature matrix
}
