% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TIL10}
\alias{TIL10}
\title{TIL10}
\format{
A data frame with 170 rows (genes) and 10 variables (cell types). The rownames contains the GeneSymbols and the colnames the cell types
}
\usage{
data(TIL10)
}
\description{
A dataset containing the molecular signatureprofiles of 10 cell types. taken from 
[Finotello al.](https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-019-0638-6)
}
\details{
\itemize{
  \item B cells
  \item Macrophages M1
  \item Macrophages M2   
  \item Monocytes
  \item Neutrophyls
  \item NK cells
  \item T cells CD4
  \item T cells CD8
  \item T cells regulatory (Tregs)
  \item Dendritic cells
}
}
\references{
\url{https://icbi-lab.github.io/immunedeconv/}
}
\keyword{datasets}
