% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TIL9}
\alias{TIL9}
\title{TIL9}
\format{
A data frame with 547 rows (genes) and 22 variables (cell types). The rownames contains the GeneSymbols and the colnames the cell types
}
\usage{
data(TIL9)
}
\description{
A dataset containing the molecular signature profiles of 9 cell types. taken from 
[Yen-Jung Chiu et al.](https://bmcmedgenomics.biomedcentral.com/articles/10.1186/s12920-019-0613-5)
\itemize{
  \item T CD4
  \item NK cells
  \item Macrophages M2   
  \item Macrophages M1   
  \item Dendritic Cells
  \item T helper
  \item T cells regulatory (Tregs)
  \item T cells CD8
  \item T cells CD8 memory
}
}
\references{

}
\keyword{datasets}
