\name{mixture}
\alias{mixture}
\docType{package}
\title{Mixture Models for Clustering and Classification}
\description{
An implementation of all 14 Gaussian parsimonious clustering models (GPCMs) for model-based clustering and model-based classification.
}
\details{
\tabular{ll}{
Package: \tab mixture \cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2015-03-10\cr
License: \tab GPL (>=2)\cr
}
This package contains the functions \code{\link{gpcm}}, \code{\link{e.step}}, and \code{\link{m.step}} as well as one simulated data set.
}
\author{
Ryan P. Browne, Aisha ElSherbiny and Paul D. McNicholas.

Maintainer: Ryan Browne  <rbrowne@math.mcmaster.ca>
}
\seealso{
Details, examples, and references are given under \code{\link{gpcm}}.
}
