\name{mixture}
\alias{mixture}
\docType{package}
\title{Mixture Models for Clustering and Classification}
\description{
An implementation of all 14 Gaussian parsimonious clustering models (GPCMs) for model-based clustering and model-based classification.
}
\details{
\tabular{ll}{
Package: \tab mixture \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-06\cr
License: \tab GPL (>=2)\cr
}
This package contains the function \code{\link{gpcm}} plus one simulated data set.
}
\author{
Ryan P. Browne and Paul D. McNicholas.

Maintainer: Ryan Browne  <rbrowne@uoguelph.ca>
}
\seealso{
Details, examples, and references are given under \code{\link{gpcm}}.
}
