% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTURE.DEBUG_V0.1.R
\name{MIXTURE}
\alias{MIXTURE}
\title{MIXTURE
the MIXTURE algorithm. It solves Y = X*B estimating the B coefficients or cell type proportions (B>0)}
\usage{
MIXTURE(
  expressionMatrix,
  signatureMatrix,
  iter = 100,
  functionMixture,
  useCores,
  verbose = FALSE,
  nullDist = c("none", "SingleSubjectTest", "PopulationBased"),
  fileSave
)
}
\arguments{
\item{expressionMatrix}{a GxS gene expression matrix. with row names genes ID as in the signature Matrix. Genes in rows, samples in columns}

\item{signatureMatrix}{a NxL gene signature matrix for L cell types}

\item{iter}{integer, number of iterations to estimate the p values}

\item{functionMixture}{: the deconvolution function: cibersort, nu.svm.robust.RFE, ls.rfe.abbas (tiped without quotation marks)}

\item{useCores}{: (integer) the number of cpus to use.}

\item{verbose}{: boolean. if TRUE, msgs on the screen}

\item{nullDist}{: (character) select the null distribution mode for p.valu estimation of the regression estimate
one of the followings : "none" , "PopulationBased" use the whole expressionMatrix to draw "iter" random samples at once. 
             Then all the models are compared against this null distribution}

\item{filesave}{: (character) the name of the output EXCEL(r) file in xlsx format}
}
\value{
an S3 object of class MIXTURE  (list) with the following slots: 
Subjects        : a list with the following slots:
       MIXabs    : a data.frame with S rows, and L columns with the absoluite regression coefficients
       MIXprop   : a data.frame with S rows, and L columns with the proportions
       ACCmetric : a data.frame with S rows, and RMSEa (Root mean squared error from absolute coeficientes),
       RMSEp (RMSE with proportions coefficientes)
       Ra (correlation absolute coeffs)
       Rp (correlation proportion coeffs)
       BestParams (Best model parameters)Iter (number of itereation of the RFE to reach the best model)
       
 PermutedMatrix  : the simulated wxpression matrix for null distribution estimation (if nullDist == "none", PermutedMatrix = NULL)
 p.values        : The p.values for each subject and for each of the ACC metrix columns (if nullDist == "none", p.value = NULL)
 method          : the null distribution method (nullDist)
 usedGenes       : the intersection gene list between expressionMatrix and signatureMatrix
}
\description{
MIXTURE
the MIXTURE algorithm. It solves Y = X*B estimating the B coefficients or cell type proportions (B>0)
}
